Apply crossbow feats when reloading using Conjure Bullet
